fix(lolWorker): Fixed an issue where some champs icon wouldn't display in embed
This issue was caused due to the fact that the champs name in-game is different to how the image on the cdn is named. For example: This champ in game is called 'FiddleSticks' but the cdn wants 'Fiddlesticks'. To resolve this I just lowercased the entire string then made the first char uppercase just how the cdn wants it.